third_party/blink/web_tests/platform/generic/accessibility/canvas-accessibilitynodeobject-expected.txt

Link Button     
1
2
 ARIA button ARIA link

This test makes sure that AccessibilityNodeObjects are created for elements in a canvas subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axLayoutObjects.length is axNodeObjects.length
PASS i == 0; axLayoutObject.role == axNodeObject.role is true
PASS i == 1; axLayoutObject.role == axNodeObject.role is true
PASS i == 2; axLayoutObject.role == axNodeObject.role is true
PASS i == 3; axLayoutObject.role == axNodeObject.role is true
PASS i == 4; axLayoutObject.role == axNodeObject.role is true
PASS i == 5; axLayoutObject.role == axNodeObject.role is true
PASS i == 6; axLayoutObject.role == axNodeObject.role is true
PASS i == 7; axLayoutObject.role == axNodeObject.role is true
PASS i == 8; axLayoutObject.role == axNodeObject.role is true
PASS i == 9; axLayoutObject.role == axNodeObject.role is true
PASS i == 10; axLayoutObject.role == axNodeObject.role is true
PASS successfullyParsed is true

TEST COMPLETE
